requirements.txt
fastapi ==0.103.1
uvicorn ==0.23.2
langchain == 0.2.3
langchain-community ==0.2.4
langchain-groq == 0.1.5
python-dotenv ==1.0.1
pdfminer.six ==20231228
sentence-transformers== 3.3.1
tiktoken ==0.7.0
semantic-text-splitter==0.19.0
chromadb == 0.5.0
